:recycle: refactor: replace NotificationFactory with NotificationDTO and remove unused factory code
